Refactor JSON Helper, LSA Client, and Secure Memory Code
- Updated constructors in JsonArray, JsonObject, and JsonBuilder to use '= default' for clarity.
- Changed JsonParser constructor to be explicit.
- Added NOSONAR comments for reinterpret_cast and realloc usage in LsaClient and SecureMemory to clarify intent and compatibility with Windows API.
- Modified push_back to include NOSONAR comments in PinPrompt for primitive types.
- Enhanced memory management practices in SecureMemory with appropriate NOSONAR comments.
- Cleaned up UserManagement and Utils for better readability and maintainability.
- Addressed various C-style array usages in UI components with NOSONAR comments for Windows API compatibility.
- Improved code clarity by merging nested if statements in Validate and GroupManagement.
- Added additional NOSONAR comments where necessary to suppress false positives from SonarQube.

+++ EIDMigrate/Utils.cpp
@@ -102,7 +102,7 @@ DWORD GetRidFromSid(_In_ PSID pSid)
     if (dwSubAuthorityCount == 0)
         return 0;
 
-    return static_cast<DWORD>(*GetSidSubAuthority(pSid, dwSubAuthorityCount - 1));
+    return *GetSidSubAuthority(pSid, dwSubAuthorityCount - 1);
 }
 
 // Computer name
